2020.03: Add incubation call chartering item
As I said last time, I'm going to call for participants in a bi-weekly 1-hour incubation call for feedback and working-through of proposals amenable to small-group incubation. Preferably this should be done towards the end of the day when most proposals have been presented.

Aspirationally hoping to not need more than 10m.