2022.03: Add ShadowRealms clarification entry
Ref https://github.com/tc39/proposal-shadowrealm/pull/348

In the TC39 Dec 2021 plenary we got consensus to set `name` and `length` for Wrapped Exotic functions within the ShadowRealms proposal. Even though, the meeting notes ended somehow ambiguous about the `name` property being prefixed with "wrapped " or not. The intention from the champions - as presented in that plenary - was to not use the prefix.