2022.06: Add ShadowRealm topics
this adds 2 new sub-topics for ShadowRealms in addition to a pre-existing one (errors across the ShadowRealm boundary).

I'm estimating the HTML Integration requiring extra time, but I'm happy to flex the proposed timebox if necessary.

There is a meta issue on https://github.com/tc39/proposal-shadowrealm/issues/365 with a more detailed list of the expected sub-topics.