2022.11: Move schedule constraints to the correct sub-section
As far as I understand, all constraints should go either in the "Normal constraints" or in the "Late-breaking schedule constraints" sections.

It's important that they are properly sorted, because as the comment in the agenda says "Conflicting constraints honored on a first-come, first served basis". This PR sorts them in the same order as they were committed (note: there are multiple commits touching Frank's constraint, I picked the oldest one).