Remove Allen's standards participation guide
I don't want us to tell new delegates, "Don’t Talk Too Much", which comes up very prominently and early in this document. Instead, we should encourage new delegates to talk! We can reference this document from some places, as it has interesting ideas, but only if we can frame it with that context.